requirements.txt
torch
torchvision
Flask
Pillow
scikit-image
